Moved PropertyTest.calcSizes to PropCheckerAsserting companion object.
